+++ frontend/src/app/page.jsx
@@ -6,20 +6,6 @@ import { useState } from 'react';
 import { useRouter } from 'next/navigation';
 
 const Home = () => {
-  // For the matching and collaboration service
-  const router = useRouter();
-
-  const [roomId, setRoomId] = useState('');
-  const [isMatched, setIsMatched] = useState(false);
-  const onMatch = (roomId) => {
-    setRoomId(roomId);
-    setIsMatched(true);
-
-    console.log("room id is routed to: " + roomId);
-    router.push(`/collab-page/${roomId}`);
-    //router.push(`/questions/6533d92691995349640128f3`);
-  };
-
   return (
     <section className="w-full flex-center flex-col">
       <h1 className="text-center text-5xl">
@@ -33,7 +19,6 @@ const Home = () => {
         Browse the questions here and get matched
       </p>
       <QuestionsList />
-      <Matching onMatch={onMatch} />
     </section>
   );
 };
